requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
git+https://github.com/openai/CLIP.git
numpy>=1.21.0
pandas>=1.5.0
pillow>=9.0.0
opencv-python>=4.8.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0
pyyaml>=6.0.0
scikit-learn>=1.2.0

# Vector database
faiss-cpu>=1.7.4
# chromadb>=0.4.0  # Commented out as it may cause conflicts

# UI
streamlit>=1.25.0
gradio>=3.35.0

# Development
jupyter>=1.0.0
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
